backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
opencv-python
numpy
scipy
scikit-learn
web3
python-dotenv
email-validator
PyJWT
passlib
bcrypt